# openapi.yml
openapi: 3.0.3
info:
  title: IQA-application
  version: 0.1.0
  description: API of IQA-application
  contact:
    name: admin
    email: admin@intocode.ru
servers:
  - url: 'https://iqa-backend.herokuapp.com'
    description: Staging server
  - url: 'http://localhost:3030'
    description: Local server
paths:
...
  /user/favorites:
    get:
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 'Вывод избранных вопросов пользователя'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionData'
        '401':
          description: 'Ошибка авторизации'
        '400':
          description: 'Неправильный, некорректный запрос'
...
  /tags:
    get:
      summary: 'Поиск по тэгам'
      parameters:
        - in: query
          name: q
          schema:
            type: string
          required: true
          description: 'Поисковая фраза'
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 'Вывод всех тэгов по поиску'
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                  $ref: '#/components/schemas/TagData'
        '400':
          description: 'Неправильный, некорректный запрос'
    post:
      security:
        - bearerAuth: []
      description: 'Добавление нового тэга в базу'
      requestBody:
        description: 'Данные при добавлении нового тэга'
        content:
          application/json:
            schema:
              properties:
                name:
                  type: string
                  example: 'nameTag'
                color:
                  type: string
                  example: 'primary'
      responses:
        '200':
          description: 'Ответ после добавления тэга в базу'
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                  $ref: '#/components/schemas/TagData'
        '401':
          description: 'Ошибка авторизации'
        '400':
          description: 'Неправильный, некорректный запрос'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: jwt
  schemas:
    UserData:
      properties:
        _id:
          type: string
          example: '61e299dfsl3a51al5261e65b'
        name:
          type: string
          example: 'IvanIvanov'
        email:
          type: string
          example: 'ivanmagimedov@mail.ru'
        githubId:
          type: string
          example: '93570249'
        avatarUrl:
          type: string
          example: 'https://avatars.githubusercontent.com/u/93570249?v=4'
        favorites:
          type: array
          items:
            type: string
            example: '61c470e91d5e851eb6fd8d47'
        createdAt:
          type: string
          example: '2022-01-15T09:54:39.531Z'
        updatedAt:
          type: string
          example: '2022-02-22T10:57:49.668Z'
        avatar:
          type: string
          example:
            thumbnail: 'https://avatars.githubusercontent.com/u/93570249?v=4&s=40'
            full: 'https://avatars.githubusercontent.com/u/93570249?v=4'
...
    QuestionData:
      properties:
        _id:
          type: string
          example: '61e299dfsl3a51al5261e65b'
        question:
          type: string
          example: 'The more your tests resemble the way your software is used, the more confidence they can give you.'
        comment:
          type: string
          example: 'You want to write maintainable tests that give you high confidence that your components are working for your users. As a part of this goal, you want your tests to avoid including implementation details so refactors of your components (changes to implementation but not functionality) do not break your tests and slow you and your team down'
        tags:
          type: array
          items:
            type: string
            $ref: '#/components/schemas/TagData'
        user:
          $ref: '#/components/schemas/UserData'
        createdAt:
          type: string
          example: '2022-01-15T09:54:39.531Z'
        updatedAt:
          type: string
          example: '2022-02-22T10:57:49.668Z'
        deleted:
          type: boolean
          example: false
        commentsCount:
          type: number
          example: 0
        questionRateCount:
          type: number
          default: 0
          example: 2
    TagData:
      properties:
        _id:
          type: string
          example: '61e299dfsl3a51al5261e65b'
        name:
          type: string
          example: jest
          minLength: 1
          maxLength: 15
        color:
          type: string
          example: 'primary'
        createdAt:
          type: string
          example: '2022-01-15T09:54:39.531Z'
        updatedAt:
          type: string
          example: '2022-02-22T10:57:49.668Z'
...